Refactor: Renamed `Outfiles` to `FileCreators`
